Split out uart drivers, mailbox, registers
